+++ crates/nu-parser/src/lib.rs
@@ -1,11 +1,13 @@
+mod errors;
 mod lite_parse;
 mod parse;
 mod path;
 mod shapes;
 mod signature;
 
-pub use crate::lite_parse::{lite_parse, LiteBlock};
-pub use crate::parse::{classify_block, garbage, parse_full_column_path};
-pub use crate::path::expand_ndots;
-pub use crate::shapes::shapes;
-pub use crate::signature::{Signature, SignatureRegistry};
+pub use errors::ParseError;
+pub use lite_parse::{lite_parse, LiteBlock};
+pub use parse::{classify_block, garbage, parse_full_column_path};
+pub use path::expand_ndots;
+pub use shapes::shapes;
+pub use signature::{Signature, SignatureRegistry};
